requirements.txt
Django==3.1
djangorestframework==3.12.2
django-test-without-migrations==0.6
Faker==8.1.2
django-environ